fix: state history id should *not* be unsigned!
What was I thinking?  It did not even save any memory to make it uint16

+++ src/state_align_search.h
@@ -58,7 +58,7 @@ extern "C" {
  * History structure
  */
 struct state_align_hist_s {
-    uint16 id;
+    int32 id;
     int32 score;
 };
 typedef struct state_align_hist_s state_align_hist_t;
